test: Generalize `createCoreResources` to global lib
